To-do: 1. Check up on how to automate Microsoft C++ Packages installations for DDS 3. Check up on how to disable privacy settings and skip OOBE process for newly created accounts 4. Follow up on account creation and deletion process in N1 Clean Manual 5. Check up on dism.exe hanging fix
On-hold:
1. Create code to retrieve device profile information and append it to a csv file

+++ N1 Clean Autopilot/Resources/Script1.ps1
@@ -120,10 +120,13 @@ Start-Process -FilePath "Comber x64 1.0.1.0.exe" -ArgumentList /v"/passive" -Wai
 $tempXmlFile = "$locAppDataTmpDir\AppAssoc.xml"
 
 # Clean up any existing folders before proceeding
-Remove-Item -Path $tempXmlFile -Force -ErrorAction SilentlyContinue
+if (Test-Path -Path $tempXmlFile) {
+
+    Remove-Item -Path $tempXmlFile -Force -ErrorAction SilentlyContinue
+}
 
 # Export current default app associations into temporary XML file
-Start-Process dism.exe -ArgumentList "/online", "/Export-DefaultAppAssociations:$tempXmlFile" -Wait -NoNewWindow
+dism.exe /online /Export-DefaultAppAssociations:$tempXmlFile | Out-Null
 
 # Read exported XML file into an XML object for modification
 $xmlContent = [xml] (Get-Content -Path $tempXmlFile)
@@ -139,7 +142,7 @@ $pdfAssociationNode.ApplicationName = "Adobe Acrobat"
 $xmlContent.Save($tempXmlFile)
 
 # Import updated default app associations from temporary XML file
-Start-Process dism.exe -ArgumentList "/online", "/Import-DefaultAppAssociations:$tempXmlFile" -Wait -NoNewWindow
+dism.exe /online /Import-DefaultAppAssociations:$tempXmlFile | Out-Null
 
 # Delete temporary XML file
 Remove-Item -Path $tempXmlFile -Force -ErrorAction SilentlyContinue
